+++ R/GeneralTransform.R
@@ -14,6 +14,8 @@
 #' @param outgoing_controlTableKeys character, which column names of the outgoing control table are considered to be keys.
 #' @param tmp_name_source a tempNameGenerator from cdata::mk_tmp_name_source()
 #' @param temporary logical, if TRUE use temporary tables
+#' @param allow_rqdatatable_in logical, if TRUE allow rqdatatable shortcutting on simple conversions.
+#' @param allow_rqdatatable_out logical, if TRUE allow rqdatatable shortcutting on simple conversions.
 #' @return processing pipeline or transformed table
 #'
 #'
@@ -82,7 +84,9 @@ convert_records <- function(table,
                             incoming_controlTableKeys = colnames(incoming_shape)[[1]],
                             outgoing_controlTableKeys = colnames(outgoing_shape)[[1]],
                             tmp_name_source = wrapr::mk_tmp_name_source("crec"),
-                            temporary = TRUE) {
+                            temporary = TRUE,
+                            allow_rqdatatable_in = FALSE,
+                            allow_rqdatatable_out = TRUE) {
   wrapr::stop_if_dot_args(substitute(list(...)), "cdata::convert_records")
   if(!is.data.frame(incoming_shape)) {
     stop("cdata::convert_records incoming_shape should be a data.frame")
@@ -99,7 +103,8 @@ convert_records <- function(table,
     strict = strict,
     controlTableKeys = incoming_controlTableKeys,
     tmp_name_source = tmp_name_source,
-    temporary = temporary)
+    temporary = temporary,
+    allow_rqdatatable = allow_rqdatatable_in)
   result <- rowrecs_to_blocks(
     row_records,
     controlTable = outgoing_shape,
@@ -109,6 +114,7 @@ convert_records <- function(table,
     controlTableKeys = outgoing_controlTableKeys,
     columnsToCopy = c(keyColumns, columnsToCopy_in),
     tmp_name_source = tmp_name_source,
-    temporary = temporary)
+    temporary = temporary,
+    allow_rqdatatable = allow_rqdatatable_out)
   result
 }