+++ R/opspec.R
@@ -15,6 +15,7 @@ NULL
 #' @param checkNames passed to rowrecs_to_blocks.
 #' @param checkKeys passed to rowrecs_to_blocks.
 #' @param strict passed to rowrecs_to_blocks.
+#' @param allow_rqdatatable logical, if TRUE allow rqdatatable shortcutting on simple conversions.
 #' @return a record specification object
 #'
 #' @examples
@@ -51,8 +52,9 @@ rowrecs_to_blocks_spec <- function(controlTable,
                                    recordKeys = character(0),
                                    controlTableKeys = colnames(controlTable)[[1]],
                                    checkNames = TRUE,
-                                   checkKeys = TRUE,
-                                   strict = FALSE) {
+                                   checkKeys = FALSE,
+                                   strict = FALSE,
+                                   allow_rqdatatable = TRUE) {
   wrapr::stop_if_dot_args(substitute(list(...)), "cdata::rowrecs_to_blocks_spec")
   controlTable <- as.data.frame(controlTable)
   rownames(controlTable) <- NULL
@@ -60,12 +62,16 @@ rowrecs_to_blocks_spec <- function(controlTable,
   if(!is.null(ck)) {
     stop(paste("cdata::rowrecs_to_blocks_spec", ck))
   }
+  if(length(intersect(recordKeys, colnames(controlTable)))>0) {
+    stop("cdata::rowrecs_to_blocks_spec recordKeys intersected control table columns")
+  }
   r <- list(controlTable = controlTable,
             recordKeys = recordKeys,
             controlTableKeys = controlTableKeys,
             checkNames = checkNames,
             checkKeys = checkKeys,
-            strict = strict)
+            strict = strict,
+            allow_rqdatatable = allow_rqdatatable)
   class(r) <- "rowrecs_to_blocks_spec"
   r
 }
@@ -82,6 +88,7 @@ rowrecs_to_blocks_spec <- function(controlTable,
 #' @param checkNames passed to blocks_to_rowrecs.
 #' @param checkKeys passed to blocks_to_rowrecs.
 #' @param strict passed to blocks_to_rowrecs.
+#' @param allow_rqdatatable logical, if TRUE allow rqdatatable shortcutting on simple conversions.
 #' @return a record specification object
 #'
 #' @examples
@@ -122,7 +129,8 @@ blocks_to_rowrecs_spec <- function(controlTable,
                                    controlTableKeys = colnames(controlTable)[[1]],
                                    checkNames = TRUE,
                                    checkKeys = TRUE,
-                                   strict = FALSE) {
+                                   strict = FALSE,
+                                   allow_rqdatatable = FALSE) {
   wrapr::stop_if_dot_args(substitute(list(...)), "cdata::blocks_to_rowrecs_spec")
   controlTable <- as.data.frame(controlTable)
   rownames(controlTable) <- NULL
@@ -130,12 +138,16 @@ blocks_to_rowrecs_spec <- function(controlTable,
   if(!is.null(ck)) {
     stop(paste("cdata::blocks_to_rowrecs_spec", ck))
   }
+  if(length(intersect(recordKeys, colnames(controlTable)))>0) {
+    stop("cdata::blocks_to_rowrecs_spec recordKeys intersected control table columns")
+  }
   r <- list(controlTable = controlTable,
             recordKeys = recordKeys,
             controlTableKeys = controlTableKeys,
             checkNames = checkNames,
             checkKeys = checkKeys,
-            strict = strict)
+            strict = strict,
+            allow_rqdatatable = allow_rqdatatable)
   class(r) <- "blocks_to_rowrecs_spec"
   r
 }
@@ -247,6 +259,9 @@ t.blocks_to_rowrecs_spec <- function(x) {
 }
 
 
+
+
+
 # use the transform spec
 
 #' Use transform spec to layout data.
@@ -323,7 +338,8 @@ layout_by.rowrecs_to_blocks_spec <- function(transform, table) {
                     columnsToCopy = transform$recordKeys,
                     checkNames = transform$checkNames,
                     checkKeys = transform$checkKeys,
-                    strict = transform$strict)
+                    strict = transform$strict,
+                    allow_rqdatatable = transform$allow_rqdatatable)
 }
 
 
@@ -363,7 +379,8 @@ layout_by.blocks_to_rowrecs_spec <- function(transform, table) {
                     controlTableKeys = transform$controlTableKeys,
                     checkNames = transform$checkNames,
                     checkKeys = transform$checkKeys,
-                    strict = transform$strict)
+                    strict = transform$strict,
+                    allow_rqdatatable = transform$allow_rqdatatable)
 }
 
 
@@ -408,7 +425,8 @@ layout_by.blocks_to_rowrecs_spec <- function(transform, table) {
                     columnsToCopy = transform$recordKeys,
                     checkNames = transform$checkNames,
                     checkKeys = transform$checkKeys,
-                    strict = transform$strict)
+                    strict = transform$strict,
+                    allow_rqdatatable = transform$allow_rqdatatable)
 }
 
 #' Factor-out (aggregate/project) block records into row records.
@@ -452,7 +470,8 @@ layout_by.blocks_to_rowrecs_spec <- function(transform, table) {
                     controlTableKeys = transform$controlTableKeys,
                     checkNames = transform$checkNames,
                     checkKeys = transform$checkKeys,
-                    strict = transform$strict)
+                    strict = transform$strict,
+                    allow_rqdatatable = transform$allow_rqdatatable)
 }
 
 
@@ -477,7 +496,8 @@ apply_right.rowrecs_to_blocks_spec <- function(pipe_left_arg,
                     columnsToCopy = transform_spec$recordKeys,
                     checkNames = transform_spec$checkNames,
                     checkKeys = transform_spec$checkKeys,
-                    strict = transform_spec$strict)
+                    strict = transform_spec$strict,
+                    allow_rqdatatable = transform_spec$allow_rqdatatable)
 }
 
 #' @export
@@ -496,7 +516,258 @@ apply_right.blocks_to_rowrecs_spec <- function(pipe_left_arg,
                     controlTableKeys = transform_spec$controlTableKeys,
                     checkNames = transform_spec$checkNames,
                     checkKeys = transform_spec$checkKeys,
-                    strict = transform_spec$strict)
+                    strict = transform_spec$strict,
+                    allow_rqdatatable = transform_spec$allow_rqdatatable)
 }
 
+
+
+
+# general spec
+
+
+#' Create a record to record spec.
+#'
+#' Create a general record to record transform specification.
+#'
+#' @param incoming_shape data.frame, definition of incoming record shape.
+#' @param outgoing_shape data.frame, defintion of outgoing record shape.
+#' @param ... not used, force later arguments to bind by name.
+#' @param recordKeys vector of columns identifying records.
+#' @param incoming_controlTableKeys character, which column names of the incoming control table are considered to be keys.
+#' @param outgoing_controlTableKeys character, which column names of the outgoing control table are considered to be keys.
+#' @param checkNames passed to rowrecs_to_blocks.
+#' @param checkKeys passed to rowrecs_to_blocks.
+#' @param strict passed to rowrecs_to_blocks.
+#' @param allow_rqdatatable_in logical, if TRUE allow rqdatatable shortcutting on simple conversions.
+#' @param allow_rqdatatable_out logical, if TRUE allow rqdatatable shortcutting on simple conversions.
+#' @return a record specification object
+#'
+#' @examples
+#'
+#'
+#' incoming_shape <- qchar_frame(
+#'   "row",  "col1", "col2", "col3" |
+#'   "row1",   v11,     v12,  v13   |
+#'   "row2",   v21,     v22,  v23   |
+#'   "row3",   v31,     v32,  v33   )
+#'
+#'
+#' outgoing_shape <- qchar_frame(
+#'   "column", "row1", "row2", "row3" |
+#'   "col1",      v11,  v21  ,  v31   |
+#'   "col2",      v12,  v22  ,  v32   |
+#'   "col3",      v13,  v23  ,  v33   )
+#'
+#' data <- build_frame(
+#'   'record_id', 'row',  'col1', 'col2', 'col3'  |
+#'   1,           'row1',  1,      2,      3      |
+#'   1,           'row2',  4,      5,      6      |
+#'   1,           'row3',  7,      8,      9      |
+#'   2,           'row1',  11,     12,     13     |
+#'   2,           'row2',  14,     15,     16     |
+#'   2,           'row3',  17,     18,     19     )
+#'
+#' print(data)
+#'
+#' layout <- layout_specification(
+#'   incoming_shape = incoming_shape,
+#'   outgoing_shape = outgoing_shape,
+#'   recordKeys = 'record_id')
+#'
+#' print(layout)
+#'
+#' data %.>% layout
+#'
+#' data %.>% layout %.>% .(t(layout))
+#'
+#' @export
+#'
+layout_specification <- function(incoming_shape,
+                                 outgoing_shape,
+                                 ...,
+                                 recordKeys = character(0),
+                                 incoming_controlTableKeys = colnames(incoming_shape)[[1]],
+                                 outgoing_controlTableKeys = colnames(outgoing_shape)[[1]],
+                                 checkNames = TRUE,
+                                 checkKeys = TRUE,
+                                 strict = FALSE,
+                                 allow_rqdatatable_in = FALSE,
+                                 allow_rqdatatable_out = TRUE) {
+  ca <- blocks_to_rowrecs_spec(incoming_shape,
+                               controlTableKeys = incoming_controlTableKeys,
+                               recordKeys = character(0),
+                               checkNames = checkNames,
+                               checkKeys = checkKeys,
+                               strict = strict,
+                               allow_rqdatatable = allow_rqdatatable_in)
+  cb <- rowrecs_to_blocks_spec(outgoing_shape,
+                               controlTableKeys = outgoing_controlTableKeys,
+                               recordKeys = character(0),
+                               checkNames = checkNames,
+                               checkKeys = checkKeys,
+                               strict = strict,
+                               allow_rqdatatable = allow_rqdatatable_out)
+  # check for some trivial cases
+  if((nrow(outgoing_shape)==1) && (length(outgoing_controlTableKeys)==0)) {
+    ra <- incoming_shape %.>% ca
+    if(isTRUE(all.equal(sort(colnames(ra)), sort(colnames(outgoing_shape))))) {
+      return(blocks_to_rowrecs_spec(incoming_shape,
+                                    controlTableKeys = incoming_controlTableKeys,
+                                    recordKeys = recordKeys,
+                                    checkNames = checkNames,
+                                    checkKeys = checkKeys,
+                                    strict = strict,
+                                    allow_rqdatatable = allow_rqdatatable_in))
+    }
+  }
+  if((nrow(incoming_shape)==1) && (length(incoming_controlTableKeys)==0)) {
+    cb_inv <- t(cb)
+    rb <- outgoing_shape %.>% cb_inv
+    if(isTRUE(all.equal(sort(colnames(rb)), sort(colnames(incoming_shape))))) {
+      return(rowrecs_to_blocks_spec(outgoing_shape,
+                                    controlTableKeys = outgoing_controlTableKeys,
+                                    recordKeys = character(0),
+                                    checkNames = checkNames,
+                                    checkKeys = checkKeys,
+                                    strict = strict,
+                                    allow_rqdatatable = allow_rqdatatable_out))
+    }
+  }
+  # check transform makes sense
+  tryCatch(
+    { incoming_shape %.>% ca %.>% cb },
+    error = function(e) {
+      stop("cdata::layout_specification, incoming and outgoing shape not compatible")
+    }
+  )
+  a <- blocks_to_rowrecs_spec(incoming_shape,
+                              controlTableKeys = incoming_controlTableKeys,
+                              recordKeys = recordKeys,
+                              checkNames = checkNames,
+                              checkKeys = checkKeys,
+                              strict = strict,
+                              allow_rqdatatable = allow_rqdatatable_in)
+  b <- rowrecs_to_blocks_spec(outgoing_shape,
+                              controlTableKeys = outgoing_controlTableKeys,
+                              recordKeys = recordKeys,
+                              checkNames = checkNames,
+                              checkKeys = checkKeys,
+                              strict = strict,
+                              allow_rqdatatable = allow_rqdatatable_out)
+  r <- list(
+    blocks_to_rowrecs_spec = a,
+    rowrecs_to_blocks_spec = b)
+  class(r) <- "cdata_general_transform_spec"
+  r
+}
+
+#' @export
+#'
+apply_right.cdata_general_transform_spec <- function(pipe_left_arg,
+                                                     pipe_right_arg,
+                                                     pipe_environment,
+                                                     left_arg_name,
+                                                     pipe_string,
+                                                     right_arg_name) {
+  table <- pipe_left_arg
+  transform_spec <- pipe_right_arg
+  blocks_to_rowrecs_spec <- transform_spec$blocks_to_rowrecs_spec
+  table <- blocks_to_rowrecs(tallTable = table,
+                             keyColumns = blocks_to_rowrecs_spec$recordKeys,
+                             controlTable = blocks_to_rowrecs_spec$controlTable,
+                             controlTableKeys = blocks_to_rowrecs_spec$controlTableKeys,
+                             checkNames = blocks_to_rowrecs_spec$checkNames,
+                             checkKeys = blocks_to_rowrecs_spec$checkKeys,
+                             strict = blocks_to_rowrecs_spec$strict,
+                             allow_rqdatatable = blocks_to_rowrecs_spec$allow_rqdatatable)
+  rowrecs_to_blocks_spec <- transform_spec$rowrecs_to_blocks_spec
+  rowrecs_to_blocks(wideTable = table,
+                    controlTable = rowrecs_to_blocks_spec$controlTable,
+                    controlTableKeys = rowrecs_to_blocks_spec$controlTableKeys,
+                    columnsToCopy = rowrecs_to_blocks_spec$recordKeys,
+                    checkNames = rowrecs_to_blocks_spec$checkNames,
+                    checkKeys = rowrecs_to_blocks_spec$checkKeys,
+                    strict = rowrecs_to_blocks_spec$strict,
+                    allow_rqdatatable = rowrecs_to_blocks_spec$allow_rqdatatable)
+}
+
+#' @export
+#'
+t.cdata_general_transform_spec <- function(x) {
+  layout_specification(incoming_shape = x$rowrecs_to_blocks_spec$controlTable,
+                       outgoing_shape = x$blocks_to_rowrecs_spec$controlTable,
+                       recordKeys = x$rowrecs_to_blocks_spec$recordKeys,
+                       incoming_controlTableKeys = x$rowrecs_to_blocks_spec$controlTableKeys,
+                       outgoing_controlTableKeys = x$blocks_to_rowrecs_spec$controlTableKeys,
+                       checkNames = x$rowrecs_to_blocks_spec$checkNames,
+                       checkKeys = x$rowrecs_to_blocks_spec$checkKeys,
+                       strict = x$rowrecs_to_blocks_spec$strict,
+                       allow_rqdatatable = x$rowrecs_to_blocks_spec$allow_rqdatatable)
+}
+
+
+#' Use transform spec to layout data.
+#'
+#' @param transform object of class blocks_to_rowrecs_spec.
+#' @param table data.frame or relop.
+#' @return re-arranged data or data reference (relop).
+#'
+#'
+#' @export
+#'
+layout_by.cdata_general_transform_spec <- function(transform, table) {
+  blocks_to_rowrecs_spec <- transform$blocks_to_rowrecs_spec
+  table <- blocks_to_rowrecs(tallTable = table,
+                             keyColumns = blocks_to_rowrecs_spec$recordKeys,
+                             controlTable = blocks_to_rowrecs_spec$controlTable,
+                             controlTableKeys = blocks_to_rowrecs_spec$controlTableKeys,
+                             checkNames = blocks_to_rowrecs_spec$checkNames,
+                             checkKeys = blocks_to_rowrecs_spec$checkKeys,
+                             strict = blocks_to_rowrecs_spec$strict,
+                             allow_rqdatatable = blocks_to_rowrecs_spec$allow_rqdatatable)
+  rowrecs_to_blocks_spec <- transform$rowrecs_to_blocks_spec
+  rowrecs_to_blocks(wideTable = table,
+                    controlTable = rowrecs_to_blocks_spec$controlTable,
+                    controlTableKeys = rowrecs_to_blocks_spec$controlTableKeys,
+                    columnsToCopy = rowrecs_to_blocks_spec$recordKeys,
+                    checkNames = rowrecs_to_blocks_spec$checkNames,
+                    checkKeys = rowrecs_to_blocks_spec$checkKeys,
+                    strict = rowrecs_to_blocks_spec$strict,
+                    allow_rqdatatable = rowrecs_to_blocks_spec$allow_rqdatatable)
+}
+
+#' @export
+#'
+format.cdata_general_transform_spec <- function(x, ...) {
+  details_in <- get_transform_details(x$blocks_to_rowrecs_spec)
+  details_out <- get_transform_details(x$rowrecs_to_blocks_spec)
+  in_record <- details_in$block_record
+  out_record <- details_out$block_record
+  in_str <- wrapr::draw_framec(in_record, unquote_cols = setdiff(colnames(in_record),
+                                                                 x$blocks_to_rowrecs_spec$controlTableKeys))
+  out_str <- wrapr::draw_framec(out_record, unquote_cols = setdiff(colnames(out_record),
+                                                                   x$rowrecs_to_blocks_spec$controlTableKeys))
+  args <- x$blocks_to_rowrecs_spec[setdiff(names(x$blocks_to_rowrecs_spec),
+                                           c("controlTable", "recordKeys", "controlTableKeys"))]
+  fmt_str <- paste0("{\n ",
+                    in_str,
+                    " in_keys <- ", wrapr::map_to_char(c(x$blocks_to_rowrecs_spec$recordKeys,
+                                                         x$blocks_to_rowrecs_spec$controlTableKeys)),
+                    "\n\n # becomes\n\n " ,
+                    out_str,
+                    " out_keys <- ", wrapr::map_to_char(c(x$rowrecs_to_blocks_spec$recordKeys,
+                                                          x$rowrecs_to_blocks_spec$controlTableKeys)),
+                    "\n\n # args: ", gsub("['\"]+", "", wrapr::map_to_char(args)),
+                    "\n}\n\n")
+  fmt_str
+}
+
+#' @export
+#'
+print.cdata_general_transform_spec <- function(x, ...) {
+  fmt_str <- format(x)
+  cat(fmt_str)
+  invisible(fmt_str)
+}