+++ NEWS.md
@@ -1,4 +1,12 @@
 
+# cdata 1.1.0 2019/04/27
+
+ * Switch to rqdatatable implementation.
+ * General transform specification.
+ * More care with factors.
+ * Update vignettes.
+ * More tests with factors and dates/times.
+
 # cdata 1.0.9 2019/04/20
 
  * "layout" commands.